+++ storages/postgres.go
@@ -2,10 +2,8 @@ package storages
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"github.com/hashicorp/go-multierror"
-	"github.com/joncrlsn/dque"
 	"github.com/ksensehq/eventnative/adapters"
 	"github.com/ksensehq/eventnative/appconfig"
 	"github.com/ksensehq/eventnative/appstatus"
@@ -14,30 +12,32 @@ import (
 	"log"
 )
 
-const eventsPerPersistedFile = 2000
 const postgresStorageType = "Postgres"
 
-//Consuming event facts, put them to https://github.com/joncrlsn/dque
-//Dequeuing and store events to Postgres in streaming mode
+//Store files to Postgres in two modes:
+//batch: (1 file = 1 transaction)
+//stream: (1 object = 1 transaction)
 type Postgres struct {
+	name            string
 	adapter         *adapters.Postgres
 	tableHelper     *TableHelper
 	schemaProcessor *schema.Processor
-	eventQueue      *dque.DQue
-}
-
-type QueuedFact struct {
-	FactBytes []byte
-}
-
-// FactBuilder creates and returns a new events.Fact.
-// This is used when we load a segment of the queue from disk.
-func QueuedFactBuilder() interface{} {
-	return &QueuedFact{}
+	eventQueue      *events.PersistentQueue
+	breakOnError    bool
 }
 
 func NewPostgres(ctx context.Context, config *adapters.DataSourceConfig, processor *schema.Processor,
-	fallbackDir, storageName string) (*Postgres, error) {
+	fallbackDir, storageName string, breakOnError, streamMode bool) (*Postgres, error) {
+	var eventQueue *events.PersistentQueue
+	if streamMode {
+		var err error
+		queueName := fmt.Sprintf("%s-%s", appconfig.Instance.ServerName, storageName)
+		eventQueue, err = events.NewPersistentQueue(queueName, fallbackDir)
+		if err != nil {
+			return nil, err
+		}
+	}
+
 	adapter, err := adapters.NewPostgres(ctx, config)
 	if err != nil {
 		return nil, err
@@ -49,77 +49,50 @@ func NewPostgres(ctx context.Context, config *adapters.DataSourceConfig, process
 		return nil, err
 	}
 
-	queueName := fmt.Sprintf("%s-%s", appconfig.Instance.ServerName, storageName)
-	queue, err := dque.NewOrOpen(queueName, fallbackDir, eventsPerPersistedFile, QueuedFactBuilder)
-	if err != nil {
-		return nil, fmt.Errorf("Error opening/creating event queue for postgres: %v", err)
-	}
-
 	monitorKeeper := NewMonitorKeeper()
-	tableHelper := NewTableHelper(adapter, monitorKeeper, redshiftStorageType)
+	tableHelper := NewTableHelper(adapter, monitorKeeper, postgresStorageType)
 
 	p := &Postgres{
+		name:            storageName,
 		adapter:         adapter,
 		tableHelper:     tableHelper,
 		schemaProcessor: processor,
-		eventQueue:      queue,
+		eventQueue:      eventQueue,
+		breakOnError:    breakOnError,
 	}
-	p.start()
+
+	if streamMode {
+		p.startStreamingConsumer()
+	}
 
 	return p, nil
 }
 
 //Consume events.Fact and enqueue it
 func (p *Postgres) Consume(fact events.Fact) {
-	p.enqueue(fact)
-}
-
-//Marshaling events.Fact to json bytes and put it to persistent queue
-func (p *Postgres) enqueue(fact events.Fact) {
-	factBytes, err := json.Marshal(fact)
-	if err != nil {
-		p.logSkippedEvent(fact, fmt.Errorf("Error marshalling events fact: %v", err))
-		return
-	}
-	if err := p.eventQueue.Enqueue(QueuedFact{FactBytes: factBytes}); err != nil {
-		p.logSkippedEvent(fact, fmt.Errorf("Error putting event fact bytes to the postgres queue: %v", err))
-		return
+	if err := p.eventQueue.Enqueue(fact); err != nil {
+		logSkippedEvent(fact, err)
 	}
 }
 
 //Run goroutine to:
 //1. read from queue
-//2. insert in postgres
-//3. if error => enqueue one more time
-func (p *Postgres) start() {
+//2. insert in Postgres
+func (p *Postgres) startStreamingConsumer() {
 	go func() {
 		for {
 			if appstatus.Instance.Idle {
 				break
 			}
-			iface, err := p.eventQueue.DequeueBlock()
+			fact, err := p.eventQueue.DequeueBlock()
 			if err != nil {
 				log.Println("Error reading event fact from postgres queue", err)
-				continue
-			}
-
-			wrappedFact, ok := iface.(QueuedFact)
-			if !ok || len(wrappedFact.FactBytes) == 0 {
-				log.Println("Warn: Dequeued object is not a QueuedFact instance or wrapped events.Fact bytes is empty")
-				continue
-			}
-
-			fact := events.Fact{}
-			err = json.Unmarshal(wrappedFact.FactBytes, &fact)
-			if err != nil {
-				log.Println("Error unmarshalling events.Fact from bytes", err)
 				continue
 			}
 
 			dataSchema, flattenObject, err := p.schemaProcessor.ProcessFact(fact)
 			if err != nil {
 				log.Printf("Unable to process object %v: %v", fact, err)
-				p.enqueue(fact)
 				continue
 			}
 
@@ -130,11 +103,51 @@ func (p *Postgres) start() {
 
 			if err := p.insert(dataSchema, flattenObject); err != nil {
 				log.Printf("Error inserting to postgres table [%s]: %v", dataSchema.Name, err)
-				p.enqueue(fact)
 				continue
 			}
 		}
 	}()
+}
+
+//Store file payload to Postgres with processing
+func (p *Postgres) Store(fileName string, payload []byte) error {
+	flatData, err := p.schemaProcessor.ProcessFilePayload(fileName, payload, p.breakOnError)
+	if err != nil {
+		return err
+	}
+
+	//process db tables & schema
+	for _, fdata := range flatData {
+		dbSchema, err := p.tableHelper.EnsureTable(fdata.DataSchema)
+		if err != nil {
+			return err
+		}
+
+		if err := p.schemaProcessor.ApplyDBTyping(dbSchema, fdata); err != nil {
+			return err
+		}
+	}
+
+	//insert all data in one transaction
+	tx, err := p.adapter.OpenTx()
+	if err != nil {
+		return fmt.Errorf("Error opening postgres transaction: %v", err)
+	}
+
+	for _, fdata := range flatData {
+		for _, object := range fdata.GetPayload() {
+			if err := p.adapter.InsertInTransaction(tx, fdata.DataSchema, object); err != nil {
+				if p.breakOnError {
+					tx.Rollback()
+					return err
+				} else {
+					log.Printf("Warn: unable to insert object %v reason: %v. This line will be skipped", object, err)
+				}
+			}
+		}
+	}
+
+	return tx.DirectCommit()
 }
 
 //insert fact in Postgres
@@ -156,13 +169,24 @@ func (p *Postgres) Close() (multiErr error) {
 	if err := p.adapter.Close(); err != nil {
 		multiErr = multierror.Append(multiErr, fmt.Errorf("Error closing postgres datasource: %v", err))
 	}
-	if err := p.eventQueue.Close(); err != nil {
-		multiErr = multierror.Append(multiErr, fmt.Errorf("Error closing postgres event queue: %v", err))
+
+	if p.eventQueue != nil {
+		if err := p.eventQueue.Close(); err != nil {
+			multiErr = multierror.Append(multiErr, fmt.Errorf("Error closing postgres event queue: %v", err))
+		}
 	}
 
 	return
 }
 
-func (p *Postgres) logSkippedEvent(fact events.Fact, err error) {
+func (p *Postgres) Name() string {
+	return p.name
+}
+
+func (p *Postgres) Type() string {
+	return postgresStorageType
+}
+
+func logSkippedEvent(fact events.Fact, err error) {
 	log.Printf("Warn: unable to enqueue object %v reason: %v. This object will be skipped", fact, err)
 }